tinyusb/msc_fat_view: Use link_tables for root entries
This change extract several FAT root entries definitions
to separate files so it's easier to see how to make
custom entries.

This feature utilized newt tool features:
- link time tables (used for root dir entries)
- pkg.whole_archive switch that forces to search whole archive
- pkg.source_files (now package files are selected separately)
